# Optimize multi-fragment unfiltering, part 1