This patch improves the execution time of the attribute unfiltering path.

```
// Current
Total read query time (array open + init state + read): 15.0957 secs
  Time to unfilter attribute tiles: 8.86625 secs
```

```
// With this patch
Total read query time (array open + init state + read): 7.32202 secs
  Time to unfilter attribute tiles: 1.80354 secs
```

The issue is that the destruction time of `forward_list` is obscenely
slow within my OSX environrment. On the Linux environment that I
originally wrote this in, the `forward_list` did not cause a performance
problem. This patch just replaces the `forward_list` with a `vector`.

I have titled this as a "part 1" because I have tentatively explored
multi-threading the `unfilter_tiles` path and have observed a speedup.
More on this later.


This improves performance on Linux as well, it's just not as significant:
```
// Current
Total read query time (array open + init state + read): 8.1482 secs
  Time to unfilter attribute tiles: 2.50095 secs
```
```
// With patch
Total read query time (array open + init state + read): 5.49465 secs
  Time to unfilter attribute tiles: 0.898189 secs
```